test(loader): remove old reference binder tests
